better interface between stats panels
